Revert "Feature/schemas"

+++ controllers/AccessController.php
@@ -1,24 +1,38 @@
 <?php
-
 namespace controllers;
 
-use models\HttpConflictException;
+use App\Models\HttpConflictException;
 use Psr\Container\ContainerInterface;
-use Slim\Psr7\Response;
-use Slim\Psr7\Request;
+use Psr\Http\Message\ResponseInterface as Response;
+use Psr\Http\Message\ServerRequestInterface as Request;
 use Slim\Exception\HttpBadRequestException;
-use models\Access;
-use models\User;
 
 
 class AccessController extends Controller
 {
-    private Access $Access;
+    private $Access;
 
     public function __construct(ContainerInterface $DIcontainer)
     {
         parent::__construct($DIcontainer);
-        $this->Access = $DIcontainer->get(Access::class);
+        $this->Access = $DIcontainer->get('Access');
+    }
+
+    public function validateAccess(Request $request, array &$data): void
+    {
+        /**
+         * Validate Access
+         * 
+         * @param array $data
+         * @throws HttpBadRequestException
+         */
+        $Validator = $this->DIcontainer->get('Validator');
+        if (isset($data['name'])) {
+            if (!$Validator->validateString($data['name'])) {
+                throw new HttpBadRequestException($request, 'Incorrect access name format; pattern: '.$Validator->clearString);
+            }
+            $data['name'] = $Validator->sanitizeString($data['name']);
+        }
     }
 
     // GET /access
@@ -77,12 +91,28 @@ class AccessController extends Controller
          * @return Response $response
          */
 
-        $data = $this->getParsedData($request);
+        $data = $this->getFrom($request, [
+            "name" => 'string',
+            "access_edit" => 'boolean',
+            "buildings_view" => 'boolean',
+            "buildings_edit" => 'boolean',
+            "logs_view" => 'boolean',
+            "logs_edit" => 'boolean',
+            "rooms_view" => 'boolean',
+            "rooms_edit" => 'boolean',
+            "reservations_access" => 'boolean',
+            "reservations_confirm" => 'boolean',
+            "reservations_edit" => 'boolean',
+            "users_edit" => 'boolean',
+            "statistics_view" => 'boolean',
+        ], true);
 
-        $data['id'] = $this->Access->create($data);
+        $this->validateAccess($request, $data);
+
+        $newID = $this->Access->create($data);
         $this->Log->create([
             'user_id' => $request->getAttribute('user_id'),
-            'message' => 'USER ' . $request->getAttribute('email') . ' CREATE access DATA ' . json_encode($data)
+            'message' => 'User ' . $request->getAttribute('email') . ' created new access class id=' . $newID . ' data:' . json_encode($data)
         ]);
         return $response->withStatus(201, "Created");
     }
@@ -115,13 +145,28 @@ class AccessController extends Controller
          * @return Response $response
          */
 
-        $data = $this->getParsedData($request);
+        $data = $this->getFrom($request, [
+            "name" => 'string',
+            "access_edit" => 'boolean',
+            "buildings_view" => 'boolean',
+            "buildings_edit" => 'boolean',
+            "logs_view" => 'boolean',
+            "logs_edit" => 'boolean',
+            "rooms_view" => 'boolean',
+            "rooms_edit" => 'boolean',
+            "reservations_access" => 'boolean',
+            "reservations_confirm" => 'boolean',
+            "reservations_edit" => 'boolean',
+            "users_edit" => 'boolean',
+            "statistics_view" => 'boolean',
+        ], false);
 
-        $this->Access->setID($args['access_id']);
-        $this->Access->update($data);
+        $this->validateAccess($request, $data);
+
+        $this->Access->update($args['access_id'], $data);
         $this->Log->create([
-            'user_id' => $request->getAttribute('user_id'),
-            'message' => 'USER ' . $request->getAttribute('email') . ' UPDATE access DATA ' . json_encode($data)
+            "user_id" => $request->getAttribute('user_id'),
+            "message" => "User " . $request->getAttribute('email') . " updated data:" . json_encode($data)
         ]);
         return $response->withStatus(204, "Updated");
     }
@@ -138,12 +183,17 @@ class AccessController extends Controller
          * 
          * @return Response $response
          */
-        $access = $this->Access->read(['id' => $args['access_id']])[0];
-        $this->Access->delete($args['access_id']);
-        $this->Log->create([
-            'user_id' => $request->getAttribute('user_id'),
-            'message' => 'USER ' . $request->getAttribute('email') . ' DELETE access DATA ' . json_encode($access)
-        ]);
+        // each user with current access have to 
+        $User = $this->DIcontainer->get('User');
+        if ($User->exist(['access_id' => $args['access_id']])) {
+            throw new HttpConflictException("Some Users stil have this access class. You can't delete it", 409); //conflict
+        } else {
+            $this->Access->delete($args['access_id']);
+            $this->Log->create([
+                "user_id" => $request->getAttribute('user_id'),
+                "message" => "User " . $request->getAttribute('email') . " deleted access id=" . $args['access_id']
+            ]);
+        }
         return $response->withStatus(204, "Deleted");
     }
 }
